pyproject.toml
[tool.poetry]
name = "fastapi-test"
version = "0.1.0"
description = ""
authors = ["Julian Hinsch <jhinsch799@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_test"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.6"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
